# chatgpt_openapi.yaml
openapi: 3.1.0
info:
  title: Tana Context Mask API
  description: An embedding-backed, graph-aware semantic context acquisition layer
    for Tana Outliner and ChatGPT Actions.
  version: 1.0.0
servers:
- url: YOUR_HTTPS_TUNNEL_URL
  description: Tana Context Mask HTTPS Tunnel Endpoint
security:
- ApiKeyAuth: []
- BearerAuth: []
paths:
  /:
    get:
      summary: Root Health & Overview
      operationId: root__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /api/v1/health:
    get:
      summary: Health Check
      operationId: health_check_api_v1_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /api/v1/context/acquire:
    post:
      summary: Acquire Context for AI Task
      description: Proactively discovers, expands, and reranks relevant background
        context from Tana knowledge graph before answering a user request.
      operationId: acquireContext
      parameters:
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextAcquisitionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextPacket'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/search:
    post:
      summary: Semantic & Hybrid Search
      description: Direct hybrid semantic and BM25 keyword search across full Tana
        corpus.
      operationId: searchNodes
      parameters:
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SemanticSearchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/nodes/{node_id}:
    get:
      summary: Inspect Single Node with Graph Lineage
      description: Returns node details, parent chain, children, references, backlinks,
        supertags, and fields.
      operationId: getNode
      parameters:
      - name: node_id
        in: path
        required: true
        schema:
          type: string
          title: Node Id
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/nodes/{node_id}/expand:
    post:
      summary: Expand Graph Around Node
      description: Traverses 1-2 hops of parent, child, reference, and backlink edges
        around a target node.
      operationId: expandNode
      parameters:
      - name: node_id
        in: path
        required: true
        schema:
          type: string
          title: Node Id
      - name: hops
        in: query
        required: false
        schema:
          type: integer
          maximum: 3
          minimum: 1
          default: 1
          title: Hops
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sync:
    post:
      summary: Trigger Mirror Sync
      description: Trigger an incremental sync via Tana Remote MCP or bootstrap from
        a local export file.
      operationId: trigger_sync_api_v1_sync_post
      parameters:
      - name: mode
        in: query
        required: false
        schema:
          type: string
          enum:
          - incremental
          - bootstrap
          default: incremental
          title: Mode
      - name: lookback_days
        in: query
        required: false
        schema:
          type: integer
          maximum: 30
          minimum: 1
          default: 1
          title: Lookback Days
      - name: export_file
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Export File
      - name: max_nodes
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Nodes
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sync/status:
    get:
      summary: Get Sync & Mirror Status
      description: Returns current node count, vector count, and last sync timestamps.
      operationId: get_sync_status_api_v1_sync_status_get
      parameters:
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/traces/{trace_id}:
    get:
      summary: Inspect Retrieval Trace
      description: Returns the exact candidate nodes, scores, decisions, and graph
        expansion paths for a trace.
      operationId: get_trace_api_v1_traces__trace_id__get
      parameters:
      - name: trace_id
        in: path
        required: true
        schema:
          type: string
          title: Trace Id
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key passed via x-api-key header
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication using your secret API key
  schemas:
    ContextAcquisitionRequest:
      properties:
        task:
          type: string
          title: Task
        query:
          anyOf:
          - type: string
          - type: 'null'
          title: Query
        scope:
          anyOf:
          - type: string
          - type: 'null'
          title: Scope
        max_nodes:
          type: integer
          title: Max Nodes
          default: 8
        include_children:
          type: boolean
          title: Include Children
          default: true
        include_references:
          type: boolean
          title: Include References
          default: true
      type: object
      required:
      - task
      title: ContextAcquisitionRequest
    ContextNode:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          default: ''
        supertags:
          items:
            type: string
          type: array
          title: Supertags
        fields:
          additionalProperties: true
          type: object
          title: Fields
        breadcrumbs:
          items:
            type: string
          type: array
          title: Breadcrumbs
        children_snippets:
          items:
            type: string
          type: array
          title: Children Snippets
        references:
          items:
            type: string
          type: array
          title: References
        backlinks:
          items:
            type: string
          type: array
          title: Backlinks
        inclusion_reason:
          type: string
          title: Inclusion Reason
          default: ''
        relevance_score:
          type: number
          title: Relevance Score
          default: 0.0
        deep_link:
          type: string
          title: Deep Link
          default: ''
      type: object
      required:
      - id
      - name
      title: ContextNode
    ContextPacket:
      properties:
        trace_id:
          type: string
          title: Trace Id
        task:
          type: string
          title: Task
        summary:
          type: string
          title: Summary
        formatted_context_markdown:
          type: string
          title: Formatted Context Markdown
        nodes:
          items:
            $ref: '#/components/schemas/ContextNode'
          type: array
          title: Nodes
        total_candidates_examined:
          type: integer
          title: Total Candidates Examined
        graph_expansion_count:
          type: integer
          title: Graph Expansion Count
        latency_ms:
          type: number
          title: Latency Ms
        timestamp:
          type: string
          title: Timestamp
      type: object
      required:
      - trace_id
      - task
      - summary
      - formatted_context_markdown
      - nodes
      - total_candidates_examined
      - graph_expansion_count
      - latency_ms
      - timestamp
      title: ContextPacket
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    SearchResponse:
      properties:
        query:
          type: string
          title: Query
        total_hits:
          type: integer
          title: Total Hits
        results:
          items:
            $ref: '#/components/schemas/SearchResultItem'
          type: array
          title: Results
        latency_ms:
          type: number
          title: Latency Ms
      type: object
      required:
      - query
      - total_hits
      - results
      - latency_ms
      title: SearchResponse
    SearchResultItem:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          default: ''
        score:
          type: number
          title: Score
        semantic_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Semantic Score
        lexical_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Lexical Score
        breadcrumbs:
          items:
            type: string
          type: array
          title: Breadcrumbs
        tags:
          items:
            type: string
          type: array
          title: Tags
        parent_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Id
        last_updated:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Updated
      type: object
      required:
      - id
      - name
      - score
      title: SearchResultItem
    SemanticSearchRequest:
      properties:
        query:
          type: string
          title: Query
        limit:
          type: integer
          title: Limit
          default: 25
        tag_filter:
          anyOf:
          - type: string
          - type: 'null'
          title: Tag Filter
        scope_node_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Scope Node Id
        hybrid:
          type: boolean
          title: Hybrid
          default: true
        alpha:
          type: number
          title: Alpha
          default: 0.7
      type: object
      required:
      - query
      title: SemanticSearchRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError